Extract NavigationRoot into separate file
Move NavigationRootItem, RootSlot, and NavigationRoot from
CoordinatorStack.swift to NavigationRoot.swift for better separation.
